refactor: 企業理念コンポーネントを移動
